raftsnap: rename "snap" to "raftsnap"
