+++ raftsnap/metrics.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package snap
+package raftsnap
 
 import "github.com/prometheus/client_golang/prometheus"
